+++ apps/docs/src/app/platform/documentation/platform-documentation.component.ts
@@ -26,7 +26,7 @@ export class PlatformDocumentationComponent extends DocumentationBaseComponent {
             { url: 'platform/info-label', name: 'Info Label' },
             { url: 'platform/input', name: 'Input' },
             { url: 'platform/input-group', name: 'Input Group' },
-            { url: 'platform/feed-input', name: 'Feed Input'},
+            { url: 'platform/feed-input', name: 'Feed Input' },
             { url: 'platform/link', name: 'Link' },
             { url: 'platform/menu', name: 'Menu' },
             { url: 'platform/menu-button', name: 'Menu Button' },
@@ -64,7 +64,8 @@ export class PlatformDocumentationComponent extends DocumentationBaseComponent {
 
         this.layouts = [
             { url: 'platform/dynamic-page', name: 'Dynamic Page' },
-            { url: 'platform/form-container', name: 'Form Container' }
+            { url: 'platform/form-container', name: 'Form Container' },
+            { url: 'platform/page-footer', name: 'Page Footer' }
         ];
 
         this.utilities = [];